app/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pymysql
python-multipart
pillow
opencv-python-headless
python-dotenv
